+++ ingenialink/ethercat/network.py
@@ -25,7 +25,7 @@ class EthercatNetwork:
 
     """
 
-    FOE_APPLICATION = {"win32": {"64bit": "FoE/win_64x/FOEUpdateFirmware.exe"}}
+    FOE_APPLICATION = {"win32": {"64bit": "FoE/win_64x/FoEUpdateFirmware.exe"}}
     FOE_ERRORS = {
         1: "Can’t read the input file.",
         2: "ECAT slave can’t reach the BOOT mode.",
@@ -52,8 +52,8 @@ class EthercatNetwork:
         Raises:
             FileNotFoundError: If the firmware file cannot be found.
             ILFirmwareLoadError: If no slave is detected.
-            ILFirmwareLoadError: If the FOE write operation is not successful.
-            NotImplementedError: If FOE is not implemented for the current OS and architecture
+            ILFirmwareLoadError: If the FoE write operation is not successful.
+            NotImplementedError: If FoE is not implemented for the current OS and architecture
 
         """
         if not os.path.isfile(fw_file):
@@ -68,7 +68,7 @@ class EthercatNetwork:
                 f" {sys_name} {arch}"
             )
         exec_path = os.path.join(str(pkg_resources.files(bin_module)), app_path)
-        logger.debug(f"Call FOE application for {sys_name}-{arch}")
+        logger.debug(f"Call FoE application for {sys_name}-{arch}")
         try:
             subprocess.run([exec_path, self.interface_name, f"{slave_id}", fw_file], check=True)
         except subprocess.CalledProcessError as e: